feat: Added vector2 operations feat: Added vector2Int refactor: Removed all GDI and windows dependencies from the high level framework architecture

+++ src/Components/ListBox.cpp
@@ -18,7 +18,7 @@ ListBox::ListBox(int x, int y, int width, int height, std::string name) : Panel(
 	trackbar.Control(this);
 	this->Panel::Add(layout);
 	layout.SetAutoExtend(true);
-	layout.SetGridColumns({ width - trackbar.GetWidth() });
+	layout.SetGridColumns({ width - (int)trackbar.GetWidth() });
 	layout.SetRowGap(1);
 	layout.SetDefaultRowSize(30);
 }
